Refactoring sequence generation with SequenceGenerator class
