+++ sequence_generator_spatial.py
@@ -76,77 +76,3 @@ def sequence_autocor(um_sequence):
                 sumcor += -1/max_val
         autocor.append(sumcor/(length - dt))
     return autocor
-
-
-
-#def training_sequence(um_sequence, batch_size, data):
-#    iterable = [iter(data[i]) for i in range(len(data))] 
-#    compteur = [0 for i in range(len(data))]
-#    train_sequence = []
-#    batch_cpt = 1
-#    list_labels = []
-#    for k in um_sequence:
-#        compteur[k] += 1
-#        if batch_cpt == 1:
-#            samples, labels = next(iterable[k])
-#            tensor_samples = samples
-#            list_labels.append(labels)
-#            
-#        elif batch_cpt <= batch_size:
-#            samples, labels = next(iterable[k])
-#            tensor_samples = torch.cat((tensor_samples, samples), 0)
-#            list_labels.append(labels)
-#        if batch_cpt == batch_size:
-#            batch_cpt = 0
-#            tensor_labels = torch.tensor(list_labels)
-#            next_batch = [tensor_samples, tensor_labels]
-#            train_sequence.append(next_batch)
-#            list_labels = []
-#        if compteur[k] == 19:
-#            compteur[k] = 0
-#            random.shuffle(data[k])
-#            iterable[k] = iter(data[k])
-#        batch_cpt += 1
-#    return train_sequence
-#data = sort_dataset.sort_MNIST()
-#
-#
-#um_sequence, rates_matrix= um_sequence_generator(0, 60000, data, 1/20, 10)
-#import matplotlib.pyplot as plt
-#um_autocor= sequence_autocor(um_sequence)
-#plt.plot(um_autocor[:200])
-#
-
-#import copy
-#
-#shuffle_um_sequence = copy.deepcopy(um_sequence)
-#random.shuffle(shuffle_um_sequence)
-#autocor_control = sequence_autocor(shuffle_um_sequence)
-#
-#plt.figure()
-#plt.plot(um_autocor[1:200], label='Utrametric')
-#plt.title("Autocorrelation")
-#plt.plot(autocor_control[1:200], label='Random')
-#plt.legend()
-#
-#
-#compteur=10*[0]
-#for k in um_sequence:
-#    compteur[k]+=1
-
-
-
-#
-#
-#rates_matrix = rates_correlation.rates_cor(data, 1, 10)
-#L=[sum(rates_matrix[i]) for i in range(10)]
-#
-#lim=0
-#randomnum = random.random()
-#for i,k in enumerate(rates_vector):
-#    lim += k 
-#    print('k',k)
-#    print('lim',lim)
-#    if randomnum <= lim:
-#        print("True")
-#next_value([0], rates)
